Elasticsearch: factor out JSON handling from stage
